Module flowr::cli::client_server

message_queue implementation of the communications between the runtime client, debug client and the runtime server and debug server.

Structs

ClientConnection

ClientConnection stores information related to the connection from a runtime client to the runtime server and is used each time a message is to be sent or received.

ServerConnection

ServerConnection store information about the server side of the client/server communications between a runtime client and a runtime server and is used each time a message needs to be sent or received.

Constants

DEBUG_SERVICE_NAME

DEBUG_SERVICE_NAME can be used to discover the debug service by name

RUNTIME_SERVICE_NAME

RUNTIME_SERVICE_NAME can be used to discover the runtime service by name

WAIT

WAIT for a message to arrive when performing a receive()

Statics

DONT_WAIT

Do NOT WAIT for a message to arrive when performing a receive()

Functions

discover_service

Try to discover a server offering a particular service by name

enable_service_discovery

Start a background thread that sends out beacons for service discovery by a client every second