Enum flowr::cli::debug_server_message::DebugServerMessage

pub enum DebugServerMessage {

    JobCompleted(Job),
    EnteringDebugger,
    ExitingDebugger,
    PriorToSendingJob(Job),
    BlockBreakpoint(Block),
    FlowUnblockBreakpoint(usize),
    DataBreakpoint(String, usize, String, Value, usize, String, String, usize),
    Panic(String, usize),
    JobError(Job),
    Deadlock(String),
    ExecutionStarted,
    ExecutionEnded,
    Error(String),
    Functions(Vec<RuntimeFunction>),
    FunctionStates((RuntimeFunction, Vec<State>)),
    SendingValue(usize, Value, usize, usize),
    OverallState(RunState),
    InputState(Input),
    OutputState(Vec<OutputConnection>),
    BlockState(Vec<Block>),
    Message(String),
    Resetting,
    WaitingForCommand(usize),
    Invalid,

}

A Message sent from the debugger to a debug_client

Variants

JobCompleted(Job)

A Job ran to completion by a function

EnteringDebugger

Entering the debugger

ExitingDebugger

The debugger/run-time is exiting

PriorToSendingJob(Job)

The run-time is about to send a Job for execution

BlockBreakpoint(Block)

A breakpoint on a Block between two functions was encountered includes: blocked_id, blocking_id, blocking_io_number

FlowUnblockBreakpoint(usize)

A breakpoint on a Flow that was busy going idle (and unblocking senders to it) was hit

DataBreakpoint(String, usize, String, Value, usize, String, String, usize)

A breakpoint on a Value being sent between two functions was encountered includes: source_process_id, output_route, value, destination_id, function_name, io_name, input_number));

Panic(String, usize)

A panic occurred executing a Flows Job - includes the output of the job that panicked

JobError(Job)

There was an error executing the Job

Deadlock(String)

A check has detected that there is a deadlock between functions impeding more execution

ExecutionStarted

Execution of the flow has started

ExecutionEnded

Execution of the flow has ended

Error(String)

An error was detected - includes: A string describing the error

Functions(Vec<RuntimeFunction>)

A list of all functions

FunctionStates((RuntimeFunction, Vec<State>))

The state of a function

SendingValue(usize, Value, usize, usize)

A value is being sent from the output of one function to the input of another includes: source_process_id, value, destination_id, input_number

OverallState(RunState)

The overall state

InputState(Input)

The state of an Input - optional values on it

OutputState(Vec<OutputConnection>)

The state of an Output - list of connections

BlockState(Vec<Block>)

One or more Blocks

Message(String)

A message for display to the user of the debug_client

Resetting

The run-time is resetting the status back to the initial state

WaitingForCommand(usize)

Debugger is blocked waiting for a command before proceeding

Invalid

Invalid - used when deserialization goes wrong