Crate flowr

#![deny(missing_docs)] flowr is the flow runner. It reads a flow Manifest produced by a flow compiler, such as `flowc`` that describes the network of collaborating functions that are executed to execute the flow.

Use flowr or flowr --help or flowr -h at the comment line to see the command line options

Modules

cli_debug_client

cli_runtime_client

Structs

Error
The Error type.

Enums

ErrorKind
The kind of an error.

Traits

ResultExt
Additional methods for Result, for easy interaction with this crate.

Functions

get_flow_args

get_matches

main

num_parallel_jobs

num_threads

parse_flow_url

run

set_lib_search_path

Type Definitions

Result
Convenient wrapper around std::Result.